requirements.txt
colorama
selenium
webdriver-manager
aiohttp
phonenumbers
google
python-whois
dnspython
